Move ActivitySources to a contract files
