+++ src/JordiAragonZaragoza.SharedKernel.Application.Contracts/ApplicationActivitySources.cs
@@ -1,0 +1,7 @@
+namespace JordiAragonZaragoza.SharedKernel.Application.Contracts
+{
+    public static class ApplicationActivitySources
+    {
+        public static string Handlers => "JordiAragonZaragoza.SharedKernel.Application.Handlers";
+    }
+}
