+++ src/JordiAragonZaragoza.SharedKernel.Infrastructure.EntityFramework/Repositories/DataModel/BaseReadRepository.cs
@@ -5,8 +5,8 @@
     using System.Threading.Tasks;
     using Ardalis.Specification.EntityFrameworkCore;
     using JordiAragonZaragoza.SharedKernel.Contracts.Repositories;
+    using JordiAragonZaragoza.SharedKernel.Infrastructure.Contracts;
     using JordiAragonZaragoza.SharedKernel.Infrastructure.EntityFramework.Context;
-    using JordiAragonZaragoza.SharedKernel.Infrastructure.Interfaces;
 
     public abstract class BaseReadRepository<TDataEntity> : RepositoryBase<TDataEntity>, ISpecificationReadRepository<TDataEntity, Guid>
         where TDataEntity : class, IDataEntity
