+++ src/JordiAragonZaragoza.SharedKernel.Infrastructure/Contracts/IDataEntity.cs
@@ -1,4 +1,4 @@
-﻿namespace JordiAragonZaragoza.SharedKernel.Infrastructure.Interfaces
+﻿namespace JordiAragonZaragoza.SharedKernel.Infrastructure.Contracts
 {
     using System;
     using JordiAragonZaragoza.SharedKernel.Contracts.Model;
