+++ src/JordiAragonZaragoza.SharedKernel.Infrastructure/Contracts/InfrastructureActivitySources.cs
@@ -1,0 +1,7 @@
+namespace JordiAragonZaragoza.SharedKernel.Infrastructure.Contracts
+{
+    public static class InfrastructureActivitySources
+    {
+        public static string EventStore => "JordiAragonZaragoza.SharedKernel.Infrastructure.EventStore";
+    }
+}
