+++ src/JordiAragonZaragoza.SharedKernel.Infrastructure/JordiAragonZaragoza.SharedKernel.Infrastructure.csproj
@@ -17,7 +17,7 @@
     </ItemGroup>
 	<ItemGroup>
 	  <ProjectReference Include="..\JordiAragonZaragoza.SharedKernel\JordiAragonZaragoza.SharedKernel.csproj" />
-	  <ProjectReference Include="..\JordiAragonZaragoza.SharedKernel.Application\JordiAragonZaragoza.SharedKernel.Application.csproj" />
+	  <ProjectReference Include="..\JordiAragonZaragoza.SharedKernel.Application.Contracts\JordiAragonZaragoza.SharedKernel.Application.Contracts.csproj" />
 	  <ProjectReference Include="..\JordiAragonZaragoza.SharedKernel.Domain\JordiAragonZaragoza.SharedKernel.Domain.csproj" />
 	</ItemGroup>
 </Project>
